+++ wsidicom/codec/encoder.py
@@ -14,9 +14,9 @@
 
 """Module with encoders for image data."""
 
+import io
 from abc import ABCMeta, abstractmethod
 from enum import Enum
-import io
 from typing import Dict, Generic, Optional, Tuple, Type, TypeVar, Union
 
 import numpy as np
@@ -25,23 +25,27 @@ from PIL.Image import Image
 from pydicom import Dataset
 from pydicom.pixel_data_handlers.util import pixel_dtype
 from pydicom.uid import (
+    JPEG2000,
     UID,
     JPEGBaseline8Bit,
     JPEGExtended12Bit,
-    JPEG2000,
     JPEGLSNearLossless,
 )
+
 from wsidicom.codec.optionals import (
     IMAGE_CODECS_AVAILABLE,
-    PYLIBJPEGRLE_AVAILABLE,
-    rle_encode_frame,
     JPEG8,
+    PYLIBJPEGLS_AVAILABLE,
+    PYLIBJPEGOPENJPEG_AVAILABLE,
+    PYLIBJPEGRLE_AVAILABLE,
     jpeg2k_encode,
     jpeg8_encode,
     jpegls_encode,
+    pylibjpeg_ls_encode,
+    pylibjpeg_openjpeg_encode,
+    rle_encode_frame,
 )
 from wsidicom.codec.rle import RleCodec
-
 from wsidicom.codec.settings import (
     Channels,
     Jpeg2kSettings,
@@ -53,6 +57,7 @@ from wsidicom.codec.settings import (
     Settings,
     Subsampling,
 )
+from wsidicom.uid import HTJPEG2000
 
 SettingsType = TypeVar("SettingsType", bound=Settings)
 
@@ -61,6 +66,7 @@ class LossyCompressionIsoStandard(Enum):
     JPEG_LOSSY = "ISO_10918_1"
     JPEG_LS_NEAR_LOSSLESS = "ISO_14495_1"
     JPEG_2000_IRREVERSIBLE = "ISO_15444_1"
+    HT_JPEG_2000_IRREVERSIBLE = "ISO_15444_15"
 
     @classmethod
     def transfer_syntax_to_iso(
@@ -72,6 +78,8 @@ class LossyCompressionIsoStandard(Enum):
             return cls.JPEG_LS_NEAR_LOSSLESS
         elif transfer_syntax == JPEG2000:
             return cls.JPEG_2000_IRREVERSIBLE
+        elif transfer_syntax == HTJPEG2000:
+            return cls.HT_JPEG_2000_IRREVERSIBLE
         return None
 
 
@@ -229,8 +237,8 @@ class Encoder(Generic[SettingsType], metaclass=ABCMeta):
         encoders_by_settings: Dict[Type[Settings], Tuple[Type[Encoder], ...]] = {
             JpegSettings: (JpegEncoder, PillowEncoder),
             JpegLosslessSettings: (JpegEncoder,),
-            JpegLsSettings: (JpegLsEncoder,),
-            Jpeg2kSettings: (Jpeg2kEncoder, PillowEncoder),
+            JpegLsSettings: (JpegLsEncoder, PyJpegLsEncoder),
+            Jpeg2kSettings: (Jpeg2kEncoder, PyLibJpegOpenJpegEncoder, PillowEncoder),
             NumpySettings: (NumpyEncoder,),
             RleSettings: (PylibjpegRleEncoder, ImageCodecsRleEncoder),
         }
@@ -264,9 +272,11 @@ class PillowEncoder(Encoder[Union[JpegSettings, Jpeg2kSettings]]):
             }
             self._format = "jpeg"
         elif isinstance(settings, Jpeg2kSettings):
+            if len(settings.levels) != 1:
+                raise ValueError("Only one level is supported.")
             self._pillow_settings = {
                 "quality_mode": "dB",
-                "quality_layers": [settings.level],
+                "quality_layers": [settings.levels[0]],
                 "irreversible": not settings.lossless,
                 "mct": settings.channels == Channels.YBR,
                 "no_jp2": True,
@@ -300,7 +310,11 @@ class PillowEncoder(Encoder[Union[JpegSettings, Jpeg2kSettings]]):
                 settings.bits == 8
                 and settings.subsampling in cls._supported_subsamplings
             )
-        return isinstance(settings, Jpeg2kSettings) and settings.bits == 8
+        return (
+            isinstance(settings, Jpeg2kSettings)
+            and settings.bits == 8
+            and len(settings.levels) == 1
+        )
 
 
 class JpegEncoder(Encoder[Union[JpegSettings, JpegLosslessSettings]]):
@@ -427,6 +441,8 @@ class Jpeg2kEncoder(Encoder[Jpeg2kSettings]):
         settings: Jpeg2kSettings
             Settings for the encoder.
         """
+        if len(settings.levels) != 1:
+            raise ValueError("Only one level is supported.")
         self._bits = settings.bits
         if settings.channels == Channels.YBR:
             self._multiple_component_transform = True
@@ -436,7 +452,7 @@ class Jpeg2kEncoder(Encoder[Jpeg2kSettings]):
             self._level = 0
             self._reversible = True
         else:
-            self._level = settings.level
+            self._level = settings.levels[0]
             self._reversible = False
         super().__init__(settings)
 
@@ -464,9 +480,7 @@ class Jpeg2kEncoder(Encoder[Jpeg2kSettings]):
 
     @classmethod
     def supports_settings(cls, settings: Settings) -> bool:
-        if not isinstance(settings, Jpeg2kSettings):
-            return False
-        return True
+        return isinstance(settings, Jpeg2kSettings) and len(settings.levels) == 1
 
 
 class NumpyEncoder(Encoder[NumpySettings]):
@@ -581,3 +595,95 @@ class PylibjpegRleEncoder(RleEncoder):
     @classmethod
     def is_available(cls) -> bool:
         return PYLIBJPEGRLE_AVAILABLE
+
+
+class PyJpegLsEncoder(Encoder[JpegLsSettings]):
+    """Encoder that uses pylibjpeg-ls to encode image."""
+
+    def __init__(
+        self,
+        settings: JpegLsSettings,
+    ) -> None:
+        """Initialize PyJpegLs (JPEG LS) encoder.
+
+        Parameters
+        ----------
+        settings: JpegLsSettings
+            Settings for the encoder.
+        """
+        if settings.channels == Channels.GRAYSCALE:
+            self._interleave_mode = 0
+        else:
+            self._interleave_mode = 2
+        super().__init__(settings)
+
+    @property
+    def lossy(self) -> bool:
+        return False
+
+    def encode(self, image: Union[Image, np.ndarray]) -> bytes:
+        if not self.is_available():
+            raise RuntimeError("Pylibjpeg-ls not available.")
+        return pylibjpeg_ls_encode(
+            np.asarray(image),
+            lossy_error=self.settings.level,
+            interleave_mode=self._interleave_mode,
+        )
+
+    @classmethod
+    def is_available(cls) -> bool:
+        return PYLIBJPEGLS_AVAILABLE
+
+    @classmethod
+    def supports_settings(cls, settings: Settings) -> bool:
+        return isinstance(settings, JpegLsSettings)
+
+
+class PyLibJpegOpenJpegEncoder(Encoder[Jpeg2kSettings]):
+    """Encoder that uses pylibjpeg-openjpeg to encode image."""
+
+    def __init__(self, settings: Jpeg2kSettings) -> None:
+        """Initialize PyLibjpeg OpenJpeg (JPEG 2000) encoder.
+
+        Parameters
+        ----------
+        settings: Jpeg2kSettings
+            Settings for the encoder.
+        """
+        self._bits = settings.bits
+        if settings.channels == Channels.YBR:
+            self._multiple_component_transform = True
+        else:
+            self._multiple_component_transform = False
+        if settings.channels == Channels.GRAYSCALE:
+            self._color_space = 2
+        else:
+            self._color_space = 1
+        if settings.lossless:
+            self._levels = None
+        else:
+            self._levels = settings.levels
+        super().__init__(settings)
+
+    @property
+    def lossy(self) -> bool:
+        return not self.settings.lossless
+
+    def encode(self, image: Union[Image, np.ndarray]) -> bytes:
+        if not self.is_available():
+            raise RuntimeError("Pylibjpeg-openjpeg not available.")
+        return pylibjpeg_openjpeg_encode(
+            np.asarray(image),
+            self.bits,
+            self._color_space,
+            self._multiple_component_transform,
+            signal_noise_ratios=list(self._levels) if self._levels else None,
+        )
+
+    @classmethod
+    def is_available(cls) -> bool:
+        return PYLIBJPEGOPENJPEG_AVAILABLE
+
+    @classmethod
+    def supports_settings(cls, settings: Settings) -> bool:
+        return isinstance(settings, Jpeg2kSettings)